refactor(#114): split load.py god module into runtime/ and scaffolding.py
Breaks the 360-line load.py into focused modules:
- scaffolding.py: build_database() facade over the three generators
- runtime/excel.py: empty_excel, create_package, dump_to_excel
- runtime/http.py: requests_post, requests_bulk_post, requests_get_all, requests_update
- runtime/sync.py: update_api_from_package, get_model (now accepts models_module arg)
- runtime/__init__.py: re-exports all runtime symbols

load.py replaced with a backward-compat shim. Generated app.py template updated
to import from fastapifromfrictionless.runtime instead of fastapifromfrictionless.load,
decoupling deployed apps from the generator toolchain.
